+++ attributes/default.rb
@@ -1,2 +1,2 @@
 # Default windows timezone setting
-default['timezone-nativex']['windows_tz'] = 'Central Standard Time'
+default['timezone-simplyadrian']['windows_tz'] = 'Central Standard Time'