+++ recipes/default.rb
@@ -1,14 +1,14 @@
 #
-# Cookbook Name:: timezone-nativex
+# Cookbook Name:: timezone-simplyadrian
 # Recipe:: default
 #
-# Copyright 2015, NativeX
+# Copyright 2015, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #
 
 if windows?
-	include_recipe 'timezone-nativex::windows'
+	include_recipe 'timezone-simplyadrian::windows'
 else
 	include_recipe 'timezone-ii'
 end