+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-timezone-nativex CHANGELOG
+timezone-simplyadrian CHANGELOG
 ==========================
 
-This file is used to list changes made in each version of the timezone-nativex cookbook.
+This file is used to list changes made in each version of the timezone-simplyadrian cookbook.
 
 1.0.0
 -----
-- [Adrian Herrera] - Production release of timezone-nativex
+- [Adrian Herrera] - Production release of timezone-simplyadrian
   - Sets timezone on windows an linux systems
 
 - - -